separate version to pkg
