+++ hw12_13_14_15_calendar/cmd/calendar/main.go
@@ -9,6 +9,7 @@ import (
 	"time"
 
 	"github.com/novikov-ai/practice-misc/hw12_13_14_15_calendar/pkg/logger"
+	"github.com/novikov-ai/practice-misc/hw12_13_14_15_calendar/pkg/utils"
 
 	"github.com/novikov-ai/practice-misc/hw12_13_14_15_calendar/internal/server/grpc"
 
@@ -30,7 +31,7 @@ func main() {
 	flag.Parse()
 
 	if flag.Arg(0) == "version" {
-		printVersion()
+		utils.PrintVersion()
 		return
 	}
