+++ hw12_13_14_15_calendar/cmd/sender/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "fmt"
+
+func main() {
+	fmt.Println("SENDER")
+}
